Add CUB200 dataset support and test case

+++ dataloaders/index_list/datasets_index_generator.py
@@ -1,9 +1,12 @@
 """Generate few-shot index list for Food101 dataset."""
 
+import argparse
 import os
 import random
 
 from datasets import load_dataset
+
+from dataloaders.datasets.cub200 import Cub200Dataset
 
 
 def class_count(labels: list[int]) -> dict[int, int]:
@@ -352,6 +355,50 @@ if not os.path.exists(f"dataloaders/index_list/{dataset_name}_index_list.txt"):
             if label == class_index
         ]
         selected_samples[class_index] = random.sample(indices, samples_per_class)
+
+    with open(f"dataloaders/index_list/{dataset_name}_index_list.txt", "w") as file:
+        for class_index, indices in selected_samples.items():
+            for index in indices:
+                file.write(f"{class_index} {index}\n")
+
+# CUB_200_2011
+dataset_name = "cub200"
+if not os.path.exists(f"dataloaders/index_list/{dataset_name}_index_list.txt"):
+    args = argparse.Namespace()
+    cub200 = Cub200Dataset("data/CUB_200_2011", args)
+
+    num_classes = len(set(cub200.labels))
+    print("Total classes", num_classes)
+
+    samples_per_class = 32
+    class_counts = class_count(cub200.labels)
+    samples_per_class = min(*class_counts.values(), samples_per_class)
+    print("Samples per class", samples_per_class)
+
+    # make sure of include the samples from FSCIL literature
+    fscil_samples = []
+    for i in range(2, 12):
+        txt_path = f"dataloaders/index_list/cub200/session_{i}.txt"
+        with open(txt_path) as f:
+            fscil_samples.extend(f.read().splitlines())
+    fscil_sample_indices = {cub200.images.index(i) for i in fscil_samples}
+
+    selected_samples = {}
+    for class_index in range(num_classes):
+        indices = [i for i, label in enumerate(cub200.labels) if label == class_index]
+        # first 5 samples from FSCIL indices
+        selected_samples[class_index] = list(
+            fscil_sample_indices.intersection(set(indices)),
+        )
+        # remove already selected samples
+        indices = list(set(indices) - fscil_sample_indices)
+        # select rest of the samples randomly
+        selected_samples[class_index].extend(
+            random.sample(
+                indices,
+                samples_per_class - len(selected_samples[class_index]),
+            ),
+        )
 
     with open(f"dataloaders/index_list/{dataset_name}_index_list.txt", "w") as file:
         for class_index, indices in selected_samples.items():
